Move parse functions to its own file
