+++ client/common/client.go
@@ -4,10 +4,8 @@ import (
 	"bufio"
 	"fmt"
 	"net"
-	"strings"
 	"time"
 	"context"
-	"strconv"
 
 	"github.com/op/go-logging"
 )
@@ -95,65 +93,7 @@ func (c *Client) receiveMessage() (string, error) {
     return msg, nil
 }
 
-func (c *Client) parseResponse(response string) error {
-	parts := strings.Split(response, "|")
-	if len(parts) < 4 {
-		return fmt.Errorf("invalid_format: %q", response)
-	}
-
-	// Validar campo LEN
-	expectedLen, err := parseLenField(parts[0])
-	if err != nil {
-		return fmt.Errorf("invalid_len_field: %w", err)
-	}
-
-	// Cuerpo real de la respuesta (sin LEN=)
-	responseBody := strings.Join(parts[1:4], "|")
-	if len(responseBody) != expectedLen {
-		return fmt.Errorf("len_mismatch: expected %d, got %d", expectedLen, len(responseBody))
-	}
-
-	// Parsear status e info
-	status, info, err := parseStatusAndInfo(parts[1], parts[2])
-	if err != nil {
-		return fmt.Errorf("invalid_status_info: %w", err)
-	}
-
-	// Logs según resultado
-	if status == "success" {
-		log.Infof("action: server_response | result: success | client_id: %v | info: %v", c.config.ID, info)
-	} else {
-		log.Errorf("action: server_response | result: failed | client_id: %v | info: %v", c.config.ID, info)
-	}
-
-	return nil
-}
-
 // ---- funciones auxiliares ----
-
-func parseLenField(field string) (int, error) {
-	parts := strings.Split(field, "=")
-	if len(parts) != 2 {
-		return 0, fmt.Errorf("bad format in LEN field: %q", field)
-	}
-	expectedLen, err := strconv.Atoi(parts[1])
-	if err != nil {
-		return 0, fmt.Errorf("LEN not an integer: %q", parts[1])
-	}
-	return expectedLen, nil
-}
-
-func parseStatusAndInfo(statusField, infoField string) (string, string, error) {
-	statusParts := strings.Split(statusField, "=")
-	infoParts := strings.Split(infoField, "=")
-
-	if len(statusParts) != 2 || len(infoParts) != 2 {
-		return "", "", fmt.Errorf("bad format in status/info fields: %q | %q", statusField, infoField)
-	}
-
-	return statusParts[1], infoParts[1], nil
-}
-
 
 func (c *Client) StartClientLoop(ctx context.Context) {
 	for msgID := 1; msgID <= c.config.LoopAmount; msgID++ {
@@ -179,9 +119,21 @@ func (c *Client) StartClientLoop(ctx context.Context) {
 			log.Infof("action: apuesta_enviada | result: success | dni: %s | numero: %s",
 				c.config.Data.Documento, c.config.Data.Numero)
 
-			if err := c.parseResponse(response); err != nil {
+			status, info, err := ParseResponse(response)
+			if err != nil {
 				log.Errorf("action: parse_response | result: failed | client_id: %v | error: %v", c.config.ID, err)
+			} else if status == "success" {
+				log.Infof("action: server_response | result: success | client_id: %v | info: %v", c.config.ID, info)
+			} else {
+				log.Errorf("action: server_response | result: failed | client_id: %v | info: %v", c.config.ID, info)
 			}
+
+			if status == "success" {
+				log.Infof("action: server_response | result: success | client_id: %v | info: %v", c.config.ID, info)
+			} else {
+				log.Errorf("action: server_response | result: failed | client_id: %v | info: %v", c.config.ID, info)
+			}
+
 
 			select {
 			case <-ctx.Done():
